merge contribution and small style tweaks (#176)
* Update navbar.njk

I noticed you were using JavaScript for the navigation, which made it hard to get the links for your site without having to first go to the page. I don't know if that's to prevent scraping, but I figured there was a way to keep the effect while using semantic HTML. 

If you're not interested in this PR, feel free to reject it. I've enjoyed your writing and site regardless. Keep up the good work.

* stylesheet tweaks

---------

Co-authored-by: Tristan Isham <73843070+tristan-nsg@users.noreply.github.com>